18NOV

+++ app/Controllers/Http/CategoriesController.ts
@@ -6,14 +6,9 @@ import CategoryValidator from 'App/Validators/CategoryValidator'
 export default class CategoriesController {
 
   //Views 
-
   public async viewList({ view }: HttpContextContract) {
     const categories = await Category.all()
     return view.render('category/categoryList', { categories: categories })
-  }
-
-  public async viewCreate({ view }: HttpContextContract) {
-    return view.render('category/categoryCreate')
   }
 
   public async viewEdit({ view, params }: HttpContextContract) {
@@ -31,36 +26,57 @@ export default class CategoriesController {
     console.log(request.body());
     const payload = await request.validate(CategoryValidator);
     await Category.create(payload);
-    response.redirect('/category/list/')
+    response.redirect('/category/categories')
   }
 
   public async apiUpdate({ params, request, response }: HttpContextContract) {
     try {
       const payload = await request.validate(CategoryValidator);
       await Category.query().where('id', params.id).update(payload)
-      response.redirect('/category/list/')
+      response.redirect('/category/categories')
     } catch (error) {
       console.log(error);
     }
   }
 
-  public async apiDelete({ response, params }: HttpContextContract) {
-    const category = await Category.find(params.id)
-    if (category) {
-      const items = await Item
-        .query()
-        .where('category', category.name)
-
-      items.forEach(async item => {        
-        item.category="Ninguna"
-       console.log(item.$attributes);
-       console.log(item.toJSON());
-        await Item.query().where('id', item.id).update(item.toJSON())
-      });
-
-      await Category.query().where('id', params.id).delete()
-      response.redirect('/category/list/')
+  public async apiDelete({ view, request, response, params }: HttpContextContract) {
+    let data = request.body();
+    if (data.name == data.confirm_name) {
+      const category = await Category.find(params.id)
+      if (category) {
+        const items = await Item
+          .query()
+          .where('category', category.name)
+        items.forEach(async item => {
+          item.category = "NA"
+          await Item.query().where('id', item.id).update(item.toJSON())
+        });
+        await Category.query().where('id', params.id).delete()
+        response.redirect('/category/categories')
+      }
+    }
+    else {
+      return view.render('shared/messages', { source: "categoryDelete", name: data.name })
     }
   }
 
+
+  // const category = await Category.find(params.id)
+  // if (category) {
+  //   const items = await Item
+  //     .query()
+  //     .where('category', category.name)
+
+  //   items.forEach(async item => {        
+  //     item.category="NA"
+  //    console.log(item.$attributes);
+  //    console.log(item.toJSON());
+  //     await Item.query().where('id', item.id).update(item.toJSON())
+  //   });
+
+  //   await Category.query().where('id', params.id).delete()
+  //   response.redirect('/category/categories')
+  // }
 }
+
+
